+++ src/vs/editor/contrib/inlineCompletions/browser/ghostTextController.ts
@@ -8,18 +8,16 @@ import { KeyCode, KeyMod } from 'vs/base/common/keyCodes';
 import { Disposable, MutableDisposable, toDisposable } from 'vs/base/common/lifecycle';
 import { firstNonWhitespaceIndex } from 'vs/base/common/strings';
 import { IActiveCodeEditor, ICodeEditor } from 'vs/editor/browser/editorBrowser';
-import { EditorAction, EditorCommand, registerEditorCommand, ServicesAccessor } from 'vs/editor/browser/editorExtensions';
+import { EditorAction, ServicesAccessor } from 'vs/editor/browser/editorExtensions';
 import { EditorOption } from 'vs/editor/common/config/editorOptions';
 import { CursorColumns } from 'vs/editor/common/core/cursorColumns';
 import { Range } from 'vs/editor/common/core/range';
 import { EditorContextKeys } from 'vs/editor/common/editorContextKeys';
-import { inlineSuggestCommitId } from 'vs/editor/contrib/inlineCompletions/browser/consts';
 import { GhostTextModel } from 'vs/editor/contrib/inlineCompletions/browser/ghostTextModel';
 import { GhostTextWidget } from 'vs/editor/contrib/inlineCompletions/browser/ghostTextWidget';
 import * as nls from 'vs/nls';
 import { ContextKeyExpr, IContextKeyService, RawContextKey } from 'vs/platform/contextkey/common/contextkey';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { KeybindingsRegistry } from 'vs/platform/keybinding/common/keybindingsRegistry';
 
 export class GhostTextController extends Disposable {
 	public static readonly inlineSuggestionVisible = new RawContextKey<boolean>('inlineSuggestionVisible', false, nls.localize('inlineSuggestionVisible', "Whether an inline suggestion is visible"));
@@ -189,39 +187,6 @@ export class ActiveGhostTextController extends Disposable {
 	}
 }
 
-const GhostTextCommand = EditorCommand.bindToContribution(GhostTextController.get);
-
-export const commitInlineSuggestionAction = new GhostTextCommand({
-	id: inlineSuggestCommitId,
-	precondition: GhostTextController.inlineSuggestionVisible,
-	handler(x) {
-		x.commit();
-		x.editor.focus();
-	}
-});
-registerEditorCommand(commitInlineSuggestionAction);
-KeybindingsRegistry.registerKeybindingRule({
-	primary: KeyCode.Tab,
-	weight: 200,
-	id: commitInlineSuggestionAction.id,
-	when: ContextKeyExpr.and(
-		commitInlineSuggestionAction.precondition,
-		EditorContextKeys.tabMovesFocus.toNegated(),
-		GhostTextController.inlineSuggestionHasIndentationLessThanTabSize
-	),
-});
-
-registerEditorCommand(new GhostTextCommand({
-	id: 'editor.action.inlineSuggest.hide',
-	precondition: GhostTextController.inlineSuggestionVisible,
-	kbOpts: {
-		weight: 100,
-		primary: KeyCode.Escape,
-	},
-	handler(x) {
-		x.hide();
-	}
-}));
 
 export class ShowNextInlineSuggestionAction extends EditorAction {
 	public static ID = 'editor.action.inlineSuggest.showNext';
